% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tofilibminer2-package.R
\docType{package}
\name{tofilibminer2-package}
\alias{tofilibminer2}
\alias{tofilibminer2-package}
\title{tofilibminer2: Explore Your R Libraries}
\description{
Provides functions for learning about your R libraries, and the packages you have installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tofi980302/tofilibminer2}
  \item Report bugs at \url{https://github.com/tofi980302/tofilibminer2/issues}
}

}
\author{
\strong{Maintainer}: Tobias Fietze \email{tobias.fietze@slu.se} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
